refactor PasswordArchiveable to avoid deprecation warnings and generally simplify the logic.  Also adds a compatibility module to deal with API changes in the rails dirty attribute handling
