+++ lib/devise-security/models/password_archivable.rb
@@ -1,9 +1,12 @@
+require_relative 'compatibility'
+
 module Devise
   module Models
     # PasswordArchivable, this depends on the DatabaseAuthenticatable module from devise
     module PasswordArchivable
       extend ActiveSupport::Concern
-      include DatabaseAuthenticatable
+      include Devise::Models::DatabaseAuthenticatable
+      include Devise::Models::Compatibility
 
       included do
         has_many :old_passwords, as: :password_archivable, dependent: :destroy
@@ -11,9 +14,7 @@ module Devise
         validate :validate_password_archive, if: :password_present?
       end
 
-      def password_present?
-        password.present?
-      end
+      delegate :present?, to: :password, prefix: true
 
       def validate_password_archive
         errors.add(:password, :taken_in_past) if will_save_change_to_encrypted_password? && password_archive_included?
@@ -36,25 +37,10 @@ module Devise
       # @return [false] if disabled or not previously used
       def password_archive_included?
         return false unless max_old_passwords > 0
-        old_passwords_including_cur_change = self.old_passwords.order(:id).reverse_order.limit(max_old_passwords).pluck(:encrypted_password)
-        old_passwords_including_cur_change << encrypted_password_was  # include most recent change in list, but don't save it yet!
+        old_passwords_including_cur_change = old_passwords.order(:id).reverse_order.limit(max_old_passwords).pluck(:encrypted_password)
+        old_passwords_including_cur_change << encrypted_password_was # include most recent change in list, but don't save it yet!
         old_passwords_including_cur_change.any? do |old_password|
           self.class.new(encrypted_password: old_password).valid_password?(password)
-        end
-      end
-
-      # for backwards compatibility with Rails < 5.1.x
-      unless Devise.activerecord51?
-        def saved_change_to_encrypted_password?
-          encrypted_password_changed?
-        end
-
-        def encrypted_password_before_last_save
-          previous_changes['encrypted_password'].try(:first)
-        end
-
-        def will_save_change_to_encrypted_password?
-          changed_attributes['encrypted_password'].present?
         end
       end
 
@@ -75,7 +61,7 @@ module Devise
       # archive the last password before save and delete all to old passwords from archive
       def archive_password
         if max_old_passwords > 0
-          old_passwords.create!(encrypted_password: encrypted_password_was) unless encrypted_password_was.blank?
+          old_passwords.create!(encrypted_password: encrypted_password_was) if encrypted_password_was.present?
           old_passwords.order(:id).reverse_order.offset(max_old_passwords).destroy_all
         else
           old_passwords.destroy_all
